+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Calling `ImageServer.readRegion(RegionRequest)` can return the wrong pixels if `RegionRequest.getPath()` is wrong (https://github.com/qupath/qupath/issues/2164) 
 * The image overview doesn't rescale for very 'tall' images (https://github.com/qupath/qupath/issues/2163)
 * Export snapshot on Linux fails if the extension is not part of the filename (https://github.com/qupath/qupath/issues/2181)
+* Training object classifier for intensity class results in `Positive: Positive` or `Negative: Negative` (https://github.com/qupath/qupath/issues/2200)
 
 ### Dependency updates
 * Bio-Formats 8.5.0